requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Optional: TensorBoard for logging
tensorboard>=2.13.0

# Jupyter Support (if running in notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
